Update fir_filter.m
The plotting is to be w/(2*pi) i.e. omega by two pi. However Octave is not following the correct convention so if we write the previous version which is mathematically correct the plotting is in by 10 form.

So instead of writing the correct format mathematically we are writing -> w/2*pi